requirements.txt
fastapi==0.115.14
uvicorn==0.34.3
python-dotenv==1.1.1

langchain==0.3.26
langchain-core==0.3.66
langchain-community==0.3.26
langchain-google-genai==2.1.5
langchain-mistralai==0.2.10
langchain-cohere==0.4.4
langchain_openai==0.3.27

openai==1.93.0
google-ai-generativelanguage==0.6.18
google-auth==2.40.3
cohere==5.15.0
httpx==0.28.1
flask==3.1.1
